requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.7
django-channels==0.7.0
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-filter==24.3
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.5.0
Faker==37.0.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
oauthlib==3.2.2
pillow==11.1.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==77.0.3
six==1.17.0
sqlparse==0.5.3
toposort==1.10
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zope.interface==7.2
